ai_python/pyproject.toml
[project]
name = "simple-agent-template"
version = "0.1.0"
description = "Minimal LangChain agent template for LangSmith/LangGraph deployment"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langchain-openai>=1.1.13",
    "langgraph>=1.0.0",
    "langgraph-cli[inmem]>=0.4.14",
    "python-dotenv>=1.0.1",
    "requests>=2.33.0",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "pytest>=8.3.5",
    "ruff>=0.8.0",
]

[tool.uv.build-backend]
module-name = "simple_agent"
source-include = ["langgraph.json", ".env.example", "Makefile", "tests/**"]
